Updated the build to target FP 10.1, from here on out tinytlf only supports FP 10.1.
Modified the layout algorithm to recreate invalid TextLines instead of removing them from the screen. Since TextLine object creation is expensive, we cache any invalid lines that don't belong on the screen anymore, and recreate them the next time we need a new line.

Added in the ColumnBreak, denoted by the <colbr/> HTML tag. This tag forces a line break, and stops layout in the current ITextContainer. It causes layout to continue in the next container, or quit if there is no next container.