Adding in a virtualization actor for tracking the virtualization of TextBlocks.
A member of the engine, ITextEngineAnalytics is a place to intelligently cache and retrieve information about the paragraphs being rendered in the TextField. The implementation uses a SparseArray to track the indicies and pixel positions of the TextBlocks, but the interface can be implemented differently if any faster methods are found.

As a result of this change, I've updated the selection algorithm to be more intelligent. I also took resoliving selection rects out of the TextEngine, it didn't belong there.