Beta release of tinytlf v2.0.
I know I should be branching and committing to that, but I want to officially get this out and into the wild.

v2.0 isn't backwards compatible. I'm explicitly deleting old and unneeded API. In many ways, this is a complete rewrite. But it's a result of six to nine months of incubating on the concepts. I'm finally motivated again.

v2.0 uses DI (via SwiftSuspenders) instead of the Facade pattern, a rewritten CSS parser instead of F*CSS, a much better layout algorithm, and much faster gestures implementation.

I've decided to consolidate, bringing in the most exciting features into the core. This may make the framework less extensible, but ultimately smaller and faster.

Much more to come.