Adds virtualization to the data parsing routine as well as the rendering routine.
Now, DOMNodes aren't visited and ContentElements/TextBlocks aren't created if they're off screen. When the containing TextRectangle is about to be rendered, then the DOMNodes are visited (and allowed to traverse potentially large XML structures), styles are processed, images are loaded, and TextBlocks and ContentElements are created.

When a TextSector exits the screen, the TextBlock and TextLines are recycled, freeing up memory and releasing the blocks and lines for reuse.

DOMNodes and ContentElements, the results of potentially expensive parsing routines, are cached and will only be regenerated when significant damage has been done to the model structure.

Adds ClosureTSF and ClosureCEF, which allow generation and manipulation of TextRectangles and ContentElements (respectively) through closures. This eliminates the need for extra classes and enables quick behavior definition and mapping.

Adds preliminary image support. This will get much better.
Adds a table stub.

Removes the TSFactory, CEFactory, a few specialty TSFactories in favor of ClosureTSF and ClosureCEF.